refactor(fetcher): split reddit family into per-fetcher modules
Mirrors the hackernews/ layout: shared client/listing in client.rs, shared
post DTO + rendering in common.rs, one fetcher per file. Drops the
single-URL fetch_listing_with_bases abstraction and widens Post.score to
i64 so heavily-downvoted posts deserialize cleanly.
